+++ R/Archive.R
@@ -4,18 +4,11 @@
 #' Container around a [data.table::data.table] which stores all performed
 #' function calls of the Objective.
 #'
-#' @section Technical details:
-#'
-#' The data is stored in a private `.data` field that contains a
-#' [data.table::data.table] which logs all performed function calls of the [Objective].
-#' This [data.table::data.table] is accessed with the public `$data()` method. New
-#' values can be added with the `$add_evals()` method. This however is usually
-#' done through the evaluation of the [OptimInstance] by the [Optimizer].
-#'
 #' @template param_codomain
 #' @template param_search_space
 #' @template param_xdt
 #' @template param_ydt
+#' @template param_store_x_domain
 #' @export
 Archive = R6Class("Archive",
   public = list(
@@ -28,11 +21,23 @@ Archive = R6Class("Archive",
     #' Codomain of objective function.
     codomain = NULL,
 
-    #' @field start_time ([POSIXct]).
+    #' @field start_time ([POSIXct])\cr
+    #' Time stamp of when the optimization started. The time is set by the
+    #' [Optimizer].
     start_time = NULL,
 
-    #' @field check_values (`logical(1)`)
+    #' @field check_values (`logical(1)`)\cr
+    #' Determines if points and results are checked for validity.
     check_values = NULL,
+
+    #' @field data ([data.table::data.table])\cr
+    #' Contains all performed [Objective] function calls.
+    data = NULL,
+
+    #' @field store_x_domain (`logical(1)`)\cr
+    #' Determines if x values, should be stored in `$data$x_domain` as list
+    #' items. The trafo will be applied if defined in `search_space`.
+    store_x_domain = NULL,
 
     #' @description
     #' Creates a new instance of this [R6][R6::R6Class] class.
@@ -40,11 +45,12 @@ Archive = R6Class("Archive",
     #' @param check_values (`logical(1)`)\cr
     #' Should x-values that are added to the archive be checked for validity?
     #' Search space that is logged into archive.
-    initialize = function(search_space, codomain, check_values = TRUE) {
+    initialize = function(search_space, codomain, check_values = TRUE, store_x_domain = TRUE) {
       self$search_space = assert_param_set(search_space)
       self$codomain = assert_param_set(codomain)
       self$check_values = assert_flag(check_values)
-      private$.data = data.table()
+      self$data = data.table()
+      self$store_x_domain = assert_flag(store_x_domain)
     },
 
     #' @description
@@ -52,22 +58,24 @@ Archive = R6Class("Archive",
     #'
     #' @param xss_trafoed (`list()`)\cr
     #' Transformed point(s) in the *domain space*.
-    add_evals = function(xdt, xss_trafoed, ydt) {
+    #' Not stored and needed if `store_x_domain = FALSE`.
+    add_evals = function(xdt, xss_trafoed = NULL, ydt) {
       assert_data_table(xdt)
       assert_data_table(ydt)
-      assert_list(xss_trafoed)
+      assert_list(xss_trafoed, null.ok = !self$store_x_domain)
       assert_data_table(ydt[, self$cols_y, with = FALSE], any.missing = FALSE)
       if (self$check_values) {
         self$search_space$assert_dt(xdt[, self$cols_x, with = FALSE])
       }
       xydt = cbind(xdt, ydt)
       assert_subset(c(self$search_space$ids(), self$codomain$ids()), colnames(xydt))
-      xydt[, "x_domain" := list(xss_trafoed)]
-      xydt[, "timestamp" := Sys.time()]
-      batch_nr = private$.data$batch_nr
-      batch_nr = if (length(batch_nr)) max(batch_nr) + 1L else 1L
-      xydt[, "batch_nr" := batch_nr]
-      private$.data = rbindlist(list(private$.data, xydt), fill = TRUE, use.names = TRUE)
+      batch_nr = self$data$batch_nr
+      if (self$store_x_domain) {
+        set(xydt, j = "x_domain", value = list(xss_trafoed))
+      }
+      set(xydt, j = "timestamp", value = Sys.time())
+      set(xydt, j = "batch_nr", value = if (length(batch_nr)) max(batch_nr) + 1L else 1L)
+      self$data = rbindlist(list(self$data, xydt), fill = TRUE, use.names = TRUE)
     },
 
     #' @description
@@ -75,22 +83,23 @@ Archive = R6Class("Archive",
     #' the solution that minimizes / maximizes the objective function.
     #' For multi-crit optimization, the Pareto set / front.
     #'
-    #' @param m (`integer()`)\cr
-    #' Take only batches `m` into account. Default is all batches.
+    #' @param batch (`integer()`)\cr
+    #' The batch number(s) to limit the best results to. Default is
+    #' all batches.
     #'
     #' @return [data.table::data.table()].
-    best = function(m = NULL) {
+    best = function(batch = NULL) {
       if (self$n_batch == 0L) {
         stop("No results stored in archive")
       }
 
-      m = if (is.null(m)) {
+      batch = if (is.null(batch)) {
         seq_len(self$n_batch)
       } else {
-        assert_integerish(m, lower = 1L, upper = self$n_batch, coerce = TRUE)
+        assert_integerish(batch, lower = 1L, upper = self$n_batch, coerce = TRUE)
       }
       batch_nr = NULL # CRAN check
-      tab = private$.data[batch_nr %in% m]
+      tab = self$data[batch_nr %in% batch]
 
       max_to_min = mult_max_to_min(self$codomain)
       if (self$codomain$length == 1L) {
@@ -106,22 +115,6 @@ Archive = R6Class("Archive",
     },
 
     #' @description
-    #' Returns a [data.table::data.table] which contains all performed
-    #' [Objective] function calls.
-    #'
-    #' @param unnest (`character()`)\cr
-    #' Set of column names for columns to unnest via [mlr3misc::unnest()].
-    #' Unnested columns are stored in separate columns instead of list-columns.
-    #'
-    #' @return [data.table::data.table()].
-    data = function(unnest = NULL) {
-      if (is.null(unnest)) {
-        return(copy(private$.data))
-      }
-      unnest(copy(private$.data), unnest, prefix = "{col}_")
-    },
-
-    #' @description
     #' Helper for print outputs.
     format = function() {
       sprintf("<%s>", class(self)[1L])
@@ -133,13 +126,13 @@ Archive = R6Class("Archive",
     #' @param ... (ignored).
     print = function() {
       catf(format(self))
-      print(private$.data)
+      print(self$data)
     },
 
     #' @description
     #' Clear all evaluation results from archive.
     clear = function() {
-      private$.data = data.table()
+      self$data = data.table()
     }
   ),
 
@@ -147,29 +140,37 @@ Archive = R6Class("Archive",
 
     #' @field n_evals (`integer(1)`)\cr
     #' Number of evaluations stored in the archive.
-    n_evals = function() nrow(private$.data),
+    n_evals = function() nrow(self$data),
 
     #' @field n_batch (`integer(1)`)\cr
     #' Number of batches stored in the archive.
     n_batch = function() {
-      if (is.null(private$.data$batch_nr)) {
+      if (is.null(self$data$batch_nr)) {
         0L
       } else {
-        max(private$.data$batch_nr)
+        max(self$data$batch_nr)
       }
     },
 
-    #' @field cols_x (`character()`).
+    #' @field cols_x (`character()`)\cr
     #' Column names of search space parameters.
     cols_x = function() self$search_space$ids(),
 
-    #' @field cols_y (`character()`).
+    #' @field cols_y (`character()`)\cr
     #' Column names of codomain parameters.
     cols_y = function() self$codomain$ids()
-    # idx_unevaled = function() private$.data$y
   ),
 
   private = list(
     .data = NULL
   )
 )
+
+#' @export
+as.data.table.Archive = function(x, ...) { # nolint
+  if (!x$store_x_domain || nrow(x$data)==0) {
+    copy(x$data)
+  } else {
+    unnest(copy(x$data), "x_domain", prefix = "{col}_")
+  }
+}