+++ R/OptimInstanceMultiCrit.R
@@ -14,6 +14,7 @@
 #' @template param_xdt
 #' @template param_ydt
 #' @template param_search_space
+#' @template param_keep_evals
 #' @export
 OptimInstanceMultiCrit = R6Class("OptimInstanceMultiCrit",
   inherit = OptimInstance,
@@ -25,8 +26,12 @@ OptimInstanceMultiCrit = R6Class("OptimInstanceMultiCrit",
     #' @param objective ([Objective]).
     #' @param terminator ([Terminator])\cr
     #' Multi-criteria terminator.
-    initialize = function(objective, search_space = NULL, terminator) {
-      super$initialize(objective, search_space, terminator)
+    #' @param check_values (`logical(1)`)\cr
+    #' Should x-values that are added to the archive be checked for validity?
+    #' Search space that is logged into archive.
+    initialize = function(objective, search_space = NULL, terminator,
+      keep_evals = "all", check_values = TRUE) {
+      super$initialize(objective, search_space, terminator, keep_evals, check_values)
     },
 
     #' @description