+++ R/OptimizerRandomSearch.R
@@ -20,8 +20,10 @@
 #' Maximum number of points to try in a batch.}
 #' }
 #'
+#' @template section_progress_bars
+#' 
 #' @source
-#' `r tools::toRd(bibentries["bergstra_2012"])`
+#' `r format_bib("bergstra_2012")`
 #'
 #' @export
 #' @template example