+++ NEWS.md
@@ -1,8 +1,22 @@
+# bbotk 0.3.0
+
+* Improved `TerminatorPerfRearched` documentation.
+* Added `check_values` flag in `OptimInstance`, `OptimInstanceSingleCrit`, 
+  `OptimInstanceMultiCrit`, `ObjectiveRFun` and `ObjectiveRFunDt`.
+* `Archive$data` is a public field now.
+* Renamed `m` parameter to `batch` in `Archive$best()`.
+* `ArchiveBest` stores no data but records the best scoring evaluation.
+* Reduced runtime if `ObjectiveRFunDt` and a domain without trafo is used.
+* `OptimizerDesignPoints` supports `ParamUty`.
+* Start values for `OptimizerCmaes` abd `OptimizerNloptr` are created randomly
+  or from center of search space.
+* `Optimizer$optimize()` supports progress bars via the package `progressr`.
+
 # bbotk 0.2.2
 
 * Removed dependency on orphaned package `bibtex`.
 * Improved documentation.
-* Adds `OptimizerCmaes` from adagio package.
+* Added `OptimizerCmaes` from adagio package.
 
 # bbotk 0.2.1
 
@@ -13,5 +27,5 @@
 
 # bbotk 0.2
 
-- First version of the Black-Box Optimization Toolkit
+* First version of the Black-Box Optimization Toolkit